Fix return percentage in Caerleon
