+++ store/profit-tree/getters.ts
@@ -1,6 +1,6 @@
 import { GetterTree } from 'vuex'
 import { TreeState } from '../typeDefs'
-import { getReturnMaterialsPercentage, isArtefactItem } from '../utils'
+import { isArtefactItem } from '../utils'
 
 export const getters: GetterTree<TreeState, {}> = {
   /**
@@ -75,12 +75,49 @@ export const getters: GetterTree<TreeState, {}> = {
     return state.features.loadingText;
   },
 
+  /**
+   * Returns the percentage of materials returned 
+   * for profile cities
+   */
   returnMaterialPercentage: (state: TreeState) => {
     const parentItem = state.currentItemInfo.parent;
+    const itemName = state.currentItemInfo.name;
     const city = state.settings.cities.craftItems;
     const useFocus = state.settings.useFocus;
 
-    return getReturnMaterialsPercentage(parentItem, city, useFocus);
+    let returnMaterialsPercentage = useFocus ? 43.5 : 15.2;
+  
+    // Keywords for the category of items that the bonus is assigned to
+    const bonus: { [key: string]: string[] } = {
+      'Caerleon': ['TOOL'],
+      'Martlock': ['AXE', 'QUARTERSTAFF', 'FROSTSTAFF', 'SHOES_PLATE', 'OFF'],
+      'Bridgewatch': ['CROSSBOW', 'DAGGER', 'CURSEDSTAFF', 'ARMOR_PLATE', 'SHOES_CLOTH'],
+      'Lymhurst': ['SWORD', 'BOW', 'ARCANESTAFF', 'HEAD_LEATHER', 'SHOES_LEATHER'],
+      'Fort Sterling': ['HAMMER', 'SPEAR', 'HOLYSTAFF', 'HEAD_PLATE', 'ARMOR_CLOTH'],
+      'Thetford': ['MACE', 'NATURESTAFF', 'FIRESTAFF', 'ARMOR_LEATHER', 'HEAD_CLOTH']
+    };
+  
+    if (!bonus[city]) {
+      return returnMaterialsPercentage;
+    }
+
+    const categories = bonus[city];
+
+    let addBonus = categories.some(keyword => {
+      return parentItem.includes(keyword);
+    });
+    
+    if (city == 'Caerleon') {
+      addBonus = categories.some(keyword => {
+        return itemName.includes(keyword);
+      });
+    }
+
+    if (addBonus) {
+      returnMaterialsPercentage = useFocus ? 47.9 : 24.8;
+    }
+
+    return returnMaterialsPercentage;
   },
   
   /**
